add info for last commit
vuplus: Use same dvbmediasink as in code.vuplus.com

The new drivers with WMV support are not working with OpenPLi images
because OpenPLi uses an implementation of dvbmediasink compartible with
Xtrend and XP drivers.

Vu is using dreambox dvbmediasink witch is incompartible with OpenPLi.

This patch creates a new dvbmediasink, same as in code.vuplus.com with
an extra patch that fixes wma sound. Also gst-ffmpeg added as RDEPEND
because drivers don't support WMA, so it is done via sofware decoding.

Thanks to macnuts, SatKiekerd and satdreamgr team for testing patches
and providing feedback.

Special thanks to devilcosta from satdreamgr for finding that gst-ffmpeg
was the culprit.

More info: http://openpli.org/forums/topic/31730-new-drivers-for-vu/