[vuplus] it913x [BUG] Enable endpoint 3 on devices with HID interface
On some USB controllers when endpoint 3 (used by HID) is not enabled
this causes a USB reset.

The original patch by Malcolm Priestley <tvboxspy@gmail.com>
at https://linuxtv.org/patch/15074/ for usb-dvb-v2.

Backported to usb-dvb by macnuts.
Many thanks to athiok for guidance.

This is a backuport for vuplus kernels which don't have the fix
present in newer kernels.

Signed-off-by: Andy Blackburn <andrew.blackburn644@btinternet.com>