[packagegroups] Add packagegroup-base-vfat As no group has yet removed vfat make it unconditional. This is less meant as an option but as a way to provide the potentionally modularized vfat support. Azbox HD is one of those boxes that have vfat/msdos-fs as a module only.
While being at it, remove the recommends for the useless AppleTalk protocol and LLC support.
They have nothing to do with what we refer to as "AppleTalk"-support.
Actually "AppleTalk" in our distros refers to the AFP server (Which is using TCP/IP), while in reality AppleTalk is a legacy network protocol and LLC is support for legacy network interfaces.