busybox/mdev.conf start bdpoll only on insertion
mdev.conf has sr0 devices rule:
 */usr/bin/bdpoll sr0 -c
Each time you insert or remove a cd or dvd, it starts a new bdpoll process.
During the base polling phase of the media mdev threats it as a partition
mount/unmount. bdpoll should be started on discovery of the device.
Also for usb devices its not uncommon to have sr1, sr2... so include those.
thx  christophecvr

I did not notice an impact to old functionality.