[python] Add 2.7.12 - !READ DESCRIPTION! Python 2.7.12 is a bugfix release to 2.7.11
It fixes quite a lot of memleaks and other nasty bugs as well as some IPv6 support.
There is also the hope that there will be some speed improvement (--with-lto).

This commit only adds the recipes for Python 2.7.12 but your images will continue to use Python 2.7.11!
In order to use this upgrade, you have to change a line in
./meta-oe/conf/distro/<your distro>.conf
from
PREFERRED_VERSION_python ?= "2.7.11"
to
PREFERRED_VERSION_python ?= "2.7.12"

For the full changelog visit
https://hg.python.org/cpython/raw-file/v2.7.12/Misc/NEWS