[openvpn] Fix (missing) autostart OpenVPN will start one instance for each valid /etc/openvpn/*.conf , so if there is none, it will start ZERO instances.
This means there is no reason not add OpenVPN to autostart, but a really good one to do:
As long as your distro properly backups OpenVPN's config, you can now even remote flash a box.
Not recommended to do so though, unless you have a second way to access the box.