[oscam-smod] Add for all distros ... ... nah, just trying to get your attention :)
I've added recipes to build
- enigma2-plugin-softcams-oscam-smod
and
- enigma2-plugin-softcams-oscam-smod-ipv4only

You are encouraged to propagate the Dual-Stack version enigma2-plugin-softcams-oscam-smod, but sadly the Dual-Stack version DOES have some problems if the server side is wrongly configured, for that reason I ALSO supply a recipe for an IPv4-only version.
Precisely, oscam lacks the code to cycle through all IPs a hostnames resolves to, so it will always only hammer the first IP it gets the hostname resolved to, also breaking multi-homed servers but mainly failing to fall-back to IPv4 if a hostname resolves to an IPv6 when it shouldn't.

It's not my work, I just merged the efforts of
- latest oscam-trunk
- gorgone's oscam-moderm webif
and
- oscam-emu
into one repository.

I've also enhanced and fixed parts of the autoconf part that was introduced in oscam-modern, so oscam-smod also has SOME of the benefits doscam has.
It can auto-detect and auto-configure the built-in card readers and some cards, so that the user doesn't have to.

Known working auto-detections:
- [DE] Astra HD+ HD01 and HD02
- [DE] Sky V13, V14 (Sat) and V23 (Unitymedia)
- [DE] Vodafone Cable cards
- [DE] Unitymedia UM01 & UM02 cards (cable)
- [AT] ORF
- [Int] Redlight cards, incl. Astra stations Dorcel and Dorcel XXX, which would normally require additional configs
- ...

I can add more auto-detections if you can provide me with a reader config that is required to make the card work and the card's ATR.

Once installed, oscam-smod WITHOUT any further config decrypts channels using all cards I could test by just inserting them into any card reader of the box, even when hot-swapping them.