[stb-kodi_17] - remove lsb from RRECOMENDS
Should enable proper install of initd-functions, ensuring that cronie, nfs, alsa and avahi work properly.
lsb does not work in the current build anyway as lsbinitscripts are not installed due to conflict with initd-initscripts.
If something else DEPENDS on lsb, initscript.bbappend shoud be modifiedm well, hacked and provide initd-functions.