Beyonwiz updates (#507)
* beyonwiz: When the apploader is updated, write it to flash

Without this, the only way to update the apploader would be via a full
USB flash.

* beyonwiz: Switch libreader default resolution to 1080p50

* [beyonwizv2] Split zip file into full USB update and EMMC partition files

Shipping both the full USB flash image and the individual partitions in
the same zip file doubles the size for no benefit. The USB updater only
uses usb_update.bin (which contains the entire flash image) and ignores
all the other files.

* beyonwizv2: Fix kernel build and versioning issues

* [beyonwiz] Remove textlcd from MACHINE_FEATURES

It conflicts with 7segment, which is actually what the hardware has

* beyonwiz: Preserve manifests when creating images

Manifests are an essential tool for tracking changes between images.