xstreamity- add arbitrary version number
People who installed this from the plugin server are unable to install adhoc updates the author provides.
This affects those who do not wish to update the image or those who uses images that do not have daily builds

Version number 2.25 has been added, but the plugin remains autorev. There is no need to keep updating the version number in the bitbake file.